bug left

+++ core/src/edu/cornell/gdiac/main/model/Penguin.java
@@ -103,9 +103,9 @@ public class Penguin extends CapsuleObstacle {
             setVX(Math.signum(getVX())*getMaxSpeed());
         } else {
             temp.set(xDir, yDir).nor();
-            forceCache.set((float) (force*temp.x*10),0f);
+            forceCache.set(force*temp.x*10,0f);
             body.applyForce(forceCache,getPosition(),true);
-            forceCache.set(0, (float) (force*temp.y*0.1f));
+            forceCache.set(0,  force*temp.y*0.1f);
             body.applyLinearImpulse(forceCache,getPosition(),true);
         }
